requirements.txt
requests==2.31.0
selenium==4.15.0
webdriver-manager==4.0.1
lxml==4.9.3
python-dotenv==1.0.0